Refactored services/util.ts into multiple files
