+++ src/rootCmds/privateCmds/deleteDapp.tsx
@@ -3,7 +3,7 @@ import { Argv } from 'yargs';
 import { App, PrettyRequest, ApiMethodLabel } from '../../ui';
 import { ArgShape, DappNameArg, UniversalArgs } from '../../cli';
 import { RootResources } from '@eximchain/dappbot-types/spec/methods';
-import { requireAuthData, fastRender, analytics, standardTrackProps } from '../../services/util';
+import { requireAuthData, fastRender, analytics, standardTrackProps } from '../../services';
 import { DeleteDapp } from '@eximchain/dappbot-types/spec/methods/private';
 
 export const command = `${RootResources.private}/deleteDapp <DappName>`;
