+++ src/services/index.ts
@@ -1,0 +1,3 @@
+export * from './util';
+export * from './commandConfig';
+export * from './truffle';
